fix(build): uint64_t, not unsigned long long, in write_cz_events
They are the same type on macOS and different ones on Linux, where uint64_t is
unsigned long -- so &m.mt_ov (uint64_t*) would not convert and every Linux job
failed to compile. The sibling encoders all say uint64_t*; this one did not.
Apple clang cannot catch it, which is the argument for saying uint64_t
everywhere rather than remembering which platform is being built.

+++ modules/core/include/TTTR.h
@@ -2345,7 +2345,7 @@ public:
      *        makes the first record of every block absolute, and the times
      *        after it jump. Null means a single self-contained block.
      */
-    void write_cz_events(FILE* fp, TTTR* tttr, unsigned long long* previous = nullptr);
+    void write_cz_events(FILE* fp, TTTR* tttr, uint64_t* previous = nullptr);
 
     /*!
      * @brief Write events from the TTTR object to a file as SM records.
